Dashboard data population and Menu creation.
For the menu, users can now preform a server check, see what server is being connected to and change the server to connect to. There is also a premilinary
help menu which is display a local html file for documenation.

For the dashboard, data is requested from the server and populated into the table. If there is an issue connecting to the server, the user is alerted via an
error dialog box.

Some additional information has been added to the config.json file for later use and a theme.json file has been created to track theme changes for the local
instance of the software.